Implement Image Importing System (#252)

+++ Source/Core/Renderer/ERS_CLASS_VisualRenderer/ERS_CLASS_ShadowMaps.cpp
@@ -188,7 +188,7 @@ void ERS_CLASS_ShadowMaps::PrioritizeDepthMaps(std::vector<ERS_STRUCT_DepthMap*>
                 }
 
                 // Ensure That Depth Maps Aren't Out Of Range
-                DepthMapIndex = std::max(DepthMapIndex, DepthMaps.size());
+                DepthMapIndex = std::max(DepthMapIndex, (unsigned int)DepthMaps.size());
 
                 // Mark For Update
                 DepthMaps[DepthMapIndex]->ToBeUpdated = true;
